+++ workspaces/intelligent-assistant/plugins/intelligent-assistant/src/hooks/index.ts
@@ -25,6 +25,7 @@ export * from './useLastOpenedConversation';
 export * from './useLightspeedDeletePermission';
 export * from './notebooks/useLightspeedNotebooksPermission';
 export * from './useLightspeedViewPermission';
+export * from './useMcpConfigureModal';
 export * from './useNotebookConversationIds';
 export * from './useDisplayModeSettings';
 export * from './notebooks/useNotebookSession';
